Rename the WP_CLI class name to CLI
